NAV-29334: Fikser rekkefølge på useEffect
