+++ src/frontend/sider/Fagsak/Behandling/BehandlingContainer.tsx
@@ -1,21 +1,43 @@
+import { useEffect } from 'react';
+
 import { useBehandlingIdParam } from '@hooks/useBehandlingIdParam';
 import { NotFound } from '@komponenter/Error/NotFound';
 import { Behandlingslinje } from '@komponenter/Saklinje/Behandlingslinje';
 import { HenleggBehandlingModal } from '@komponenter/Saklinje/Meny/HenleggBehandling/HenleggBehandlingModal';
 import { HenleggBehandlingVeivalgModal } from '@komponenter/Saklinje/Meny/HenleggBehandling/HenleggBehandlingVeivalgModal';
+import { type SideId, sider } from '@sider/Fagsak/Behandling/Sider/sider';
+import { hentSideHref } from '@utils/miljø';
 import { Outlet } from 'react-router';
 
 import { Box, GlobalAlert, HStack, VStack } from '@navikt/ds-react';
 import { RessursStatus } from '@navikt/familie-typer';
 
 import Styles from './BehandlingContainer.module.css';
-import { BehandlingProvider } from './context/BehandlingContext';
+import { BehandlingProvider, useBehandlingContext } from './context/BehandlingContext';
 import { useHentOgSettBehandlingContext } from './context/HentOgSettBehandlingContext';
 import { Høyremeny } from './Høyremeny/Høyremeny';
 import { TabContextProvider } from './Høyremeny/TabContextProvider';
 import { TotrinnskontrollModalContextProvider } from './Høyremeny/Totrinnskontroll/TotrinnskontrollModalContextProvider';
 import { KorrigerEtterbetalingModal } from './Sider/Vedtak/KorrigerEtterbetaling/KorrigerEtterbetalingModal';
 import { Venstremeny } from './Venstremeny/Venstremeny';
+
+/**
+ * Dette er gjort fordi useEffecten her må kjøre etter at useEffectene i BehandlingProvider er kjørt. Ellers visker
+ * den tilstand og viser masse "undefined" i GUI. Dette skjer spesielt hvis man er inne på vedtakssiden f.eks. og
+ * refresher siden. Burde skrive en bedre logikk.
+ */
+export function OutletContainer() {
+    const { leggTilBesøktSide } = useBehandlingContext();
+    const sidevisning = hentSideHref(location.pathname);
+
+    useEffect(() => {
+        if (sidevisning) {
+            leggTilBesøktSide(Object.entries(sider).find(([_, side]) => side.href === sidevisning)?.[0] as SideId);
+        }
+    }, [sidevisning]);
+
+    return <Outlet />;
+}
 
 export function BehandlingContainer() {
     const { behandlingRessurs } = useHentOgSettBehandlingContext();
@@ -40,7 +62,7 @@ export function BehandlingContainer() {
                                 <Venstremeny />
                             </Box>
                             <Box className={Styles.midtreKolonne}>
-                                <Outlet />
+                                <OutletContainer />
                             </Box>
                             <Box className={Styles.høyreKolonne}>
                                 <TotrinnskontrollModalContextProvider>
